Refactor: feedback from code review
