+++ include/Common.h
@@ -37,7 +37,7 @@
 namespace fs = std::filesystem;
 
 #include "TConsole.h"
-#include "TSettings.h"
+#include "Settings.h"
 
 struct Version {
     uint8_t major;
@@ -59,32 +59,7 @@ using SparseArray = std::unordered_map<size_t, T>;
 class Application final {
 public:
     // types
-    struct TSettings {
-        std::string ServerName { "BeamMP Server" };
-        std::string ServerDesc { "BeamMP Default Description" };
-        std::string ServerTags { "Freeroam" };
-        std::string Resource { "Resources" };
-        std::string MapName { "/levels/gridmap_v2/info.json" };
-        std::string Key {};
-        std::string Password {};
-        std::string SSLKeyPath { "./.ssl/HttpServer/key.pem" };
-        std::string SSLCertPath { "./.ssl/HttpServer/cert.pem" };
-        bool HTTPServerEnabled { false };
-        int MaxPlayers { 8 };
-        bool Private { true };
-        int MaxCars { 1 };
-        bool DebugModeEnabled { false };
-        int Port { 30814 };
-        std::string CustomIP {};
-        bool LogChat { true };
-        bool SendErrors { true };
-        bool SendErrorsMessageEnabled { true };
-        int HTTPServerPort { 8080 };
-        std::string HTTPServerIP { "127.0.0.1" };
-        bool HTTPServerUseSSL { false };
-        bool HideUpdateMessages { false };
-        [[nodiscard]] bool HasCustomIP() const { return !CustomIP.empty(); }
-    };
+
 
     using TShutdownHandler = std::function<void()>;
